Mod 1 Lab 1 Ex 2 Inst need to identify simulations
The option is titled Tutorials and Simulations.  The simulations are not numbered and there are more than 7 of them. Need better instructions here.  List each of the simulated attacks you wish the lab candidate to perform.  The instructions do not match the options available - could not find the graph option for the incident for Document drops backdoor.